--- a/bm/export_test.go
+++ b/bm/export_test.go
@@ -1,1 +1,0 @@
-package bm