This PR resolves #5 .
Since lukechampine/randMap seem to be wrong with Go 1.11+, I decided not use randMap.

The following is the bench mark result for comparing our new implementation to the old one on Go 1.10 environment.
As you see, it actually speed up Team Draft Multileaving 🤣 🤣 🤣 🤣 

## without randMap
```
inputRankingNum: 2, inputRankingItemNum: 10, interleavedRankingItemNum: 5
BenchmarkMultileaving/[[0-th_bench_on_Team_Draft_Multileaving]]-2         	 1000000	      1162 ns/op

inputRankingNum: 2, inputRankingItemNum: 200, interleavedRankingItemNum: 50
BenchmarkMultileaving/[[1-th_bench_on_Team_Draft_Multileaving]]-2                              	   50000	     22018 ns/op

inputRankingNum: 2, inputRankingItemNum: 200, interleavedRankingItemNum: 200
BenchmarkMultileaving/[[2-th_bench_on_Team_Draft_Multileaving]]-2                              	   20000	     98496 ns/op

inputRankingNum: 2, inputRankingItemNum: 1000, interleavedRankingItemNum: 200
BenchmarkMultileaving/[[3-th_bench_on_Team_Draft_Multileaving]]-2                              	   20000	     72706 ns/op

inputRankingNum: 10, inputRankingItemNum: 10, interleavedRankingItemNum: 5
BenchmarkMultileaving/[[0-th_bench_on_Team_Draft_Multileaving]]#01-2                           	  500000	      2206 ns/op

inputRankingNum: 10, inputRankingItemNum: 200, interleavedRankingItemNum: 50
BenchmarkMultileaving/[[1-th_bench_on_Team_Draft_Multileaving]]#01-2                           	   50000	     22845 ns/op

inputRankingNum: 10, inputRankingItemNum: 200, interleavedRankingItemNum: 200
BenchmarkMultileaving/[[2-th_bench_on_Team_Draft_Multileaving]]#01-2                           	   10000	    126418 ns/op

inputRankingNum: 10, inputRankingItemNum: 1000, interleavedRankingItemNum: 200
BenchmarkMultileaving/[[3-th_bench_on_Team_Draft_Multileaving]]#01-2                           	   10000	    119141 ns/op

inputRankingNum: 50, inputRankingItemNum: 10, interleavedRankingItemNum: 5
BenchmarkMultileaving/[[0-th_bench_on_Team_Draft_Multileaving]]#02-2                           	  100000	     13028 ns/op

inputRankingNum: 50, inputRankingItemNum: 200, interleavedRankingItemNum: 50
BenchmarkMultileaving/[[1-th_bench_on_Team_Draft_Multileaving]]#02-2                           	   30000	     43127 ns/op

inputRankingNum: 50, inputRankingItemNum: 200, interleavedRankingItemNum: 200
BenchmarkMultileaving/[[2-th_bench_on_Team_Draft_Multileaving]]#02-2                           	   10000	    166655 ns/op

inputRankingNum: 50, inputRankingItemNum: 1000, interleavedRankingItemNum: 200
BenchmarkMultileaving/[[3-th_bench_on_Team_Draft_Multileaving]]#02-2                           	   10000	    182647 ns/op

inputRankingNum: 100, inputRankingItemNum: 10, interleavedRankingItemNum: 5
BenchmarkMultileaving/[[0-th_bench_on_Team_Draft_Multileaving]]#03-2                           	  100000	     22664 ns/op

inputRankingNum: 100, inputRankingItemNum: 200, interleavedRankingItemNum: 50
BenchmarkMultileaving/[[1-th_bench_on_Team_Draft_Multileaving]]#03-2                           	   20000	     78914 ns/op

inputRankingNum: 100, inputRankingItemNum: 200, interleavedRankingItemNum: 200
BenchmarkMultileaving/[[2-th_bench_on_Team_Draft_Multileaving]]#03-2                           	   10000	    205680 ns/op


inputRankingNum: 100, inputRankingItemNum: 1000, interleavedRankingItemNum: 200
BenchmarkMultileaving/[[3-th_bench_on_Team_Draft_Multileaving]]#03-2                           	    5000	    277190 ns/op
```

## with randMap
```
inputRankingNum: 2, inputRankingItemNum: 10, interleavedRankingItemNum: 5
BenchmarkMultileaving/[[0-th_bench_on_Team_Draft_Multileaving]]-2         	  200000	     10515 ns/op

inputRankingNum: 2, inputRankingItemNum: 200, interleavedRankingItemNum: 50
BenchmarkMultileaving/[[1-th_bench_on_Team_Draft_Multileaving]]-2                              	   20000	     90303 ns/op

inputRankingNum: 2, inputRankingItemNum: 200, interleavedRankingItemNum: 200
BenchmarkMultileaving/[[2-th_bench_on_Team_Draft_Multileaving]]-2                              	    5000	    373869 ns/op

inputRankingNum: 2, inputRankingItemNum: 1000, interleavedRankingItemNum: 200
BenchmarkMultileaving/[[3-th_bench_on_Team_Draft_Multileaving]]-2                              	    3000	    366139 ns/op

inputRankingNum: 10, inputRankingItemNum: 10, interleavedRankingItemNum: 5
BenchmarkMultileaving/[[0-th_bench_on_Team_Draft_Multileaving]]#01-2                           	  100000	     18770 ns/op

inputRankingNum: 10, inputRankingItemNum: 200, interleavedRankingItemNum: 50
BenchmarkMultileaving/[[1-th_bench_on_Team_Draft_Multileaving]]#01-2                           	   10000	    153734 ns/op

inputRankingNum: 10, inputRankingItemNum: 200, interleavedRankingItemNum: 200
BenchmarkMultileaving/[[2-th_bench_on_Team_Draft_Multileaving]]#01-2                           	    2000	    662568 ns/op

inputRankingNum: 10, inputRankingItemNum: 1000, interleavedRankingItemNum: 200
BenchmarkMultileaving/[[3-th_bench_on_Team_Draft_Multileaving]]#01-2                           	    2000	    590230 ns/op

inputRankingNum: 50, inputRankingItemNum: 10, interleavedRankingItemNum: 5
BenchmarkMultileaving/[[0-th_bench_on_Team_Draft_Multileaving]]#02-2                           	   30000	     48804 ns/op

inputRankingNum: 50, inputRankingItemNum: 200, interleavedRankingItemNum: 50
BenchmarkMultileaving/[[1-th_bench_on_Team_Draft_Multileaving]]#02-2                           	    5000	    275238 ns/op

inputRankingNum: 50, inputRankingItemNum: 200, interleavedRankingItemNum: 200
BenchmarkMultileaving/[[2-th_bench_on_Team_Draft_Multileaving]]#02-2                           	    2000	   1157164 ns/op

inputRankingNum: 50, inputRankingItemNum: 1000, interleavedRankingItemNum: 200
BenchmarkMultileaving/[[3-th_bench_on_Team_Draft_Multileaving]]#02-2                           	    1000	   1290535 ns/op

inputRankingNum: 100, inputRankingItemNum: 10, interleavedRankingItemNum: 5
BenchmarkMultileaving/[[0-th_bench_on_Team_Draft_Multileaving]]#03-2                           	   30000	     56903 ns/op

inputRankingNum: 100, inputRankingItemNum: 200, interleavedRankingItemNum: 50
BenchmarkMultileaving/[[1-th_bench_on_Team_Draft_Multileaving]]#03-2                           	    3000	    375085 ns/op

inputRankingNum: 100, inputRankingItemNum: 200, interleavedRankingItemNum: 200
BenchmarkMultileaving/[[2-th_bench_on_Team_Draft_Multileaving]]#03-2                           	    1000	   1111393 ns/op

inputRankingNum: 100, inputRankingItemNum: 1000, interleavedRankingItemNum: 200
BenchmarkMultileaving/[[3-th_bench_on_Team_Draft_Multileaving]]#03-2                           	    2000	   1234889 ns/op
```